IGNITE-24424 Move dir prefixes and filters to NodeFileTree
